+++ resources/views/welcome.blade.php
@@ -3,19 +3,11 @@
 @section('content')
 
 <div class="container-fluid">
-    {{-- <div class="row"> --}}
-        {{-- <img src="img/logo1.png" class="col-4 img-fluid"> --}}
-
-    {{-- </div> --}}
-
     <div class="row">
         <section class="col-12 order-2 col-md-4 order-md-1">
             <img src="img/1.jpg" class="img-fluid">
             <img src="img/2.jpg" class="img-fluid mt-4">
             <img src="img/3.jpg" class="img-fluid mt-4">
-            {{-- <img src="img/1.jpg" class="img-fluid">
-            <img src="img/2.jpg" class="img-fluid">
-            <img src="img/3.jpg" class="img-fluid"> --}}
         </section>
 
         <div class="col-12 order-1 col-md-8 order-md-2">
@@ -29,9 +21,6 @@
                 <p>Welcome to the HPDS Bible. Here you will find the word of the Lord brought to you in the English, King James Version. Along with full
                     text of the Bible, you will find additional Bible study materials to better understand the most important historical document in existence.
                 </p>
-               @can ('admin')
-                echo "you're admin";
-               @endcan
             </section>
 
         </div>